Worked with cosine_sim function
Jonathan and I tried to edit the cosine_sim function to handle more than one design problem that could have been added to the dataframe. I eventually decided that we might get better clustering results with just one design problem at the end of the dataset (more patterns in each cluster instead of having a cluster with a bunch of similarly-worded problems).