Tested with shorter design problem
I ran the experiment again with a very brief design problem description from sourcemaking.com. For k-means, the design problem had a cluster all to itself. The f-score for fuzzy c-means was 0.667, with builder correctly assigned to cluster 2 and the design problem (the solution was builder) assigned to cluster 0 (the closest match in the cosine similarity matrix was the visitor pattern).