Corrected correct_category values
For the Wikipedia behavioral pattern design problems I created, I had accidentally assigned them a correct_category value of 2 instead of 0, so I corrected that. Observer is still the most recommended pattern for the observer problem, and this problem and the observer pattern are assigned to the same cluster for all clustering algorithms except for the PAM algorithms.